Move Navigate to Prayer Request Page
